backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "FastAPI backend with SQLModel and PostgreSQL"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.115.12",
    "sqlalchemy==2.0.41",
    "alembic==1.16.1",
    "pydantic==2.11.5",
    "pydantic-settings==2.6.1",
    "sqlmodel==0.0.24",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.20",
    "faker==33.1.3",
    "factory-boy==3.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "ruff==0.7.4",
    "black==24.10.0",
    "mypy==1.13.0",
    "types-python-jose==3.3.4.20240106",
    "types-passlib==1.7.7.20240819",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C90", # mccabe
    "UP",  # pyupgrade
]
ignore = []
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".pytest_cache",
    "alembic",
]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "sqlalchemy", "sqlmodel"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | alembic
  )/
)
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["alembic", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
